oprava warnig
git-svn-id: file:///data/svn/enum/ccReg_server/trunk@1578 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ pqsql.cc
@@ -180,7 +180,7 @@ else
 
 len =  PQescapeString( str,  String ,  length);
 
-LOG( SQL_LOG , "escape len  %ld [%s]" , len  , str   );
+LOG( SQL_LOG , "escape len  %d [%s]" ,  (int ) len  , str   );
 
 return true;
 }
